[test]랭크 테스트 코드 구현 완료 close #167

+++ src/main/java/io/pp/arcade/domain/rank/dto/RankUserDto.java
@@ -16,7 +16,7 @@ public class RankUserDto {
     private String statusMessage;
     private double winRate;
 
-    public static RankUserDto from (RankRedis userRank, Long rank){
+    public static RankUserDto from (RankRedis userRank, Integer rank){
          RankUserDto dto;
         /*
          if (userRank == null) {
